+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The client here will eventually be released as "spython" (and eventually to
 singularity on pypi), and the versions here will coincide with these releases.
 
 ## [master](https://github.com/singularityhub/singularity-cli/tree/master)
+ - client is not honoring quiet for pull (0.0.71)
  - removing debugging line in pull (0.0.70)
  - adding quiet argument to build to override client (0.0.69)
  - adding additional options to build to support singularity-compose (0.0.68)